Reworked to support the new jobs. Still a bit ugly, but all of the existing unit tests pass. Need to have the new job tests/conversions work now.
